src/pyproject.toml
[project]
name = "parlor"
version = "0.1.0"
description = "On-device, real-time multimodal AI — have natural voice + vision conversations, fully local and private"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.135.3",
    "kokoro-onnx>=0.5.0; sys_platform == 'linux' or sys_platform == 'win32'",
    "openai>=1.0.0",
    "misaki[en]>=0.9.4; sys_platform == 'darwin'",
    "mlx-audio>=0.4.2; sys_platform == 'darwin'",
    "num2words>=0.5.14; sys_platform == 'darwin'",
    "huggingface-hub>=0.20.0",
    "numpy>=2.4.4",
    "pillow>=11.0.0",
    "soundfile>=0.13.1",
    "uvicorn[standard]>=0.43.0",
    "python-dotenv>=1.0.0",
    "websockets>=16.0",
]
